backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
slowapi==0.1.9
sqlalchemy==2.0.40
alembic==1.15.2
pydantic[email]==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.11.0
twilio==9.6.1
openai==1.75.0
pytesseract==0.3.13
Pillow==11.2.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.0
httpx==0.28.1
python-dotenv==1.1.0
